# docs/openapi/paella-openapi-full.yaml
openapi: 3.0.0
info:
  title: 'Billie - PaD API Documentation'
  description: |
    # Introduction
    
    With PayAfterDelivery (PAD) / "Rechnungskauf" Billie offers a convenient, secure and simple way to allow your B2B-customers
    to pay on invoice. Our solution enables merchants to accept invoice payments from other business customers without waiting for their money,
    reminding customers about the payment, or assuming credit risk.
    
    The following documentation describes our RESTFul API, which allows easy integration into your web-shop,
    eCommerce platform or invoicing back-end.
    
    Based upon your Billie contract, your customers will have between 7 and 120 days (payment term) to pay your invoices, while Billie will 
    transfer the money to you immediately after shipment of goods or fulfilment of service obligation. For outstanding payments, 
    Billie takes over reminding customers on your behalf (white-label solution). Additionally, Billie protects you from credit risks 
    by absorbing any potential debtor defaults.
    
    ---
    
    ## Order States
    
    | State         | Description                                                                                   |
    |---------------|-----------------------------------------------------------------------------------------------|
    | waiting       | The order creation failed and needs a manual operation in order to be approved or declined    |
...
    
    ## Order State Transitions
    
...
servers:
...
paths:
...
  /internal-docs:
    get:
      tags:
        - 'API Docs'
      summary: 'Internal API Docs (default)'
      operationId: docs_internal_get_default
      parameters:
        -
          name: embed
          in: query
          description: 'Use inline JSON instead of the external file'
          required: false
          schema:
            type: boolean
          example: 1
        -
          name: nocache
          in: query
          description: 'Do not send cache headers to the client'
          required: false
          schema:
            type: boolean
          example: 1
      responses:
        '200':
          $ref: '#/components/responses/HtmlDocument'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
      x-groups:
        - private
...
  /docs:
    get:
      tags:
        - 'API Docs'
      summary: 'Public API Docs (default)'
      operationId: docs_get_default
      parameters:
        -
          name: embed
          in: query
          description: 'Use inline JSON instead of the external file'
          required: false
          schema:
            type: boolean
          example: 1
        -
          name: nocache
          in: query
          description: 'Do not send cache headers to the client'
          required: false
          schema:
            type: boolean
          example: 1
      responses:
        '200':
          $ref: '#/components/responses/HtmlDocument'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
      x-groups:
        - private
...
  /merchant:
    post:
      tags:
        - Merchants
      summary: 'Create Merchant'
      operationId: merchant_create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMerchantRequest'
      responses:
        '201':
          description: 'Merchant successfully created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateMerchantResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: 'Request conflicts with the current state of the server.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorsObject'
        '500':
          $ref: '#/components/responses/ServerError'
      x-groups:
        - private
  '/order/{uuid}/add-invoice':
    post:
      tags:
        - Support
      summary: 'Create Order Invoice'
      description: '(Automated) Creates a new invoice, linking an order with a file. Called automatically by the invoice uploader Lambda service.'
      operationId: order_create_invoice
      parameters:
        -
          name: uuid
          in: path
          description: 'Order UUID'
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - file_id
                - invoice_number
              properties:
                file_id:
                  description: 'File ID in the Nachos file service.'
                  type: integer
                invoice_number:
                  description: 'Invoice number provided by the merchant.'
                  $ref: '#/components/schemas/TinyText'
              type: object
      responses:
        '201':
          description: 'Order invoice successfully created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
      x-groups:
        - private
...
  '/merchant-debtor/{merchantDebtorUuid}/synchronize':
    post:
      tags:
        - Support
      summary: 'Synchronize merchant_debtor address.'
      operationId: synchronize_merchant_debtor
      parameters:
        -
          name: Uuid
          in: path
          description: 'Mechant Debtor UUID'
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: 'Debtor synchronized successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantDebtorSynchronizationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
      x-groups:
        - private
...
  '/checkout-session/{sessionUuid}/confirm':
    put:
      tags:
        - 'Checkout Server'
      summary: 'Checkout Session Confirm'
      description: 'Confirms the creation of an existing checkout session.'
      operationId: checkout_session_confirm
      parameters:
        -
          name: sessionUuid
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutSessionConfirmOrderRequest'
      responses:
        '202':
          description: 'Order data successfully confirmed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
      security:
        -
          oauth2: []
      x-groups:
        - public
        - private
...
  /legal-forms:
    get:
      tags:
        - 'Back-end Order Creation'
      summary: 'Get Legal Forms'
      operationId: get_legal_forms
      responses:
        '200':
          description: 'List of supported legal forms'
          content:
            application/json:
              schema:
                properties:
                  items: { type: array, items: { properties: { code: { description: 'Legal form code', type: integer, example: '90101' }, name: { description: 'Legal form name', type: string, example: 'AG & Co. KG', nullable: true }, required_input: { description: 'Required input (HR-NR => Registration Number, Ust-ID => Registration ID)', type: string, example: HR-NR }, required: { description: Required, type: boolean } }, type: object } }
                type: object
        '500':
          $ref: '#/components/responses/ServerError'
      security: []
      x-groups:
        - public
        - private
...
components:
  schemas:
    UUID:
      type: string
      format: uuid
      maxLength: 36
      minLength: 36
      example: 12345667-890a-bcde-f123-34567890abcd
    OrderState:
      title: 'Order State'
      type: string
      enum:
        - new
        - authorized
        - waiting
        - created
        - declined
        - shipped
        - paid_out
        - late
        - complete
        - canceled
        - pre_approved
      example: created
    OrderDeclineReason:
      title: 'Order Decline Reason'
      type: string
      enum:
        - risk_policy
        - debtor_not_identified
        - debtor_address
        - debtor_limit_exceeded
      example: debtor_not_identified
    OrderDunningStatus:
      title: 'Order Dunning Status'
      type: string
      enum:
        - not_started
        - active
        - paused
        - inactive
      example: active
      nullable: true
    OrderDuration:
      title: 'Order Duration'
      type: integer
      maximum: 120
      minimum: 1
      example: 30
    PostalCode:
      title: 'Postal Code'
      type: string
      maxLength: 5
      example: '10969'
...
    DateTime:
      title: Date-Time
      type: string
      format: date-time
      example: '2019-01-20 14:00:00'
    TinyText:
      title: string
      type: string
      maxLength: 255
...
    CheckoutSessionConfirmOrderRequest:
      required:
        - amount
        - duration
      properties:
        amount:
          $ref: '#/components/schemas/CreateOrderAmountRequest'
        duration:
          $ref: '#/components/schemas/OrderDuration'
      type: object
    CreateMerchantRequest:
      title: 'Merchant Data'
      properties:
        company_id:
          type: string
          format: numeric
        merchant_financing_limit:
          type: number
          format: float
        initial_debtor_financing_limit:
          type: number
          format: float
        debtor_financing_limit:
          type: number
          format: float
        webhook_url:
          type: string
          format: uri
          nullable: true
        webhook_authorization:
          description: 'Authorization header that will be sent with the merchant webhooks. Currently `X-Api-Key: XXX` and `Authorization: Basic XXX` are supported.'
          type: string
          example: 'X-Api-Key: test'
          nullable: true
      type: object
    CreateMerchantResponse:
      properties:
        oauth_client_id:
          $ref: '#/components/schemas/UUID'
        oauth_client_secret:
          $ref: '#/components/schemas/UUID'
      type: object
      allOf:
        -
          $ref: '#/components/schemas/MerchantEntity'
...
    CreateOrderAddressRequest:
      title: Address
      required:
        - street
        - city
        - postal_code
        - country
      properties:
        addition:
          nullable: true
          $ref: '#/components/schemas/TinyText'
        house_number:
          example: '4'
          nullable: true
          $ref: '#/components/schemas/TinyText'
        street:
          example: 'Charlotten Str.'
          $ref: '#/components/schemas/TinyText'
        city:
          example: Berlin
          $ref: '#/components/schemas/TinyText'
        postal_code:
          $ref: '#/components/schemas/PostalCode'
        country:
          $ref: '#/components/schemas/CountryCode'
      type: object
    CreateOrderAmountRequest:
      title: 'Order Amount'
      required:
        - net
        - gross
        - tax
      properties:
        net:
          type: number
          format: float
          minimum: 1
          example: 100
        gross:
          type: number
          format: float
          minimum: 1
          example: 119
        tax:
          type: number
          format: float
          minimum: 0
          example: 19
      type: object
...
    MerchantEntity:
      properties:
        id:
          type: integer
        created_at:
          $ref: '#/components/schemas/DateTime'
        updated_at:
          $ref: '#/components/schemas/DateTime'
        name:
          $ref: '#/components/schemas/TinyText'
        available_financing_limit:
          type: number
          format: float
        api_key:
          $ref: '#/components/schemas/TinyText'
        company_id:
          $ref: '#/components/schemas/TinyText'
        payment_merchant_id:
          $ref: '#/components/schemas/UUID'
        is_active:
          type: boolean
        webhook_url:
          type: string
          format: uri
          nullable: true
        webhook_authorization:
          type: string
          example: 'Authorization: Basic test'
          nullable: true
      type: object
...
    MerchantDebtorSynchronizationResponse:
      title: 'Merchant Debtor Synchronization'
      properties:
        debtor_company:
          description: 'Updated debtor company data'
          properties:
            name:
              type: string
              example: 'Some company'
            address_house_number:
              type: string
              example: '123'
            address_street:
              type: string
              example: 'Random street'
            address_postal_code:
              type: string
              example: '10356'
            address_city:
              type: string
              example: Berlin
            address_country:
              type: string
              example: Germany
          type: object
        is_updated:
          type: boolean
          example: 'true'
      type: object
...
    OrderResponse:
      title: 'Order Entity'
      properties:
        order_id:
          type: string
          example: C-10123456789-0001
          nullable: true
        uuid:
          $ref: '#/components/schemas/UUID'
        state:
          example: created
          $ref: '#/components/schemas/OrderState'
        reasons:
          type: string
          enum:
            - risk_policy
            - debtor_not_identified
            - debtor_address
            - debtor_limit_exceeded
          nullable: true
          deprecated: true
        decline_reason:
          nullable: true
          $ref: '#/components/schemas/OrderDeclineReason'
        amount:
          type: number
          format: float
          example: 123.45
          nullable: false
        amount_tax:
          type: number
          format: float
          example: 123.45
          nullable: false
        amount_net:
          type: number
          format: float
          example: 123.45
          nullable: false
        duration:
          example: 30
          $ref: '#/components/schemas/OrderDuration'
        dunning_status:
          $ref: '#/components/schemas/OrderDunningStatus'
        debtor_company:
          description: 'Identified company'
          properties:
            name:
              example: 'Billie GmbH'
              nullable: true
              $ref: '#/components/schemas/TinyText'
            address_house_number:
              example: '4'
              nullable: true
              $ref: '#/components/schemas/TinyText'
            address_street:
              example: Charlottenstr.
              nullable: true
              $ref: '#/components/schemas/TinyText'
            address_postal_code:
              type: string
              maxLength: 5
              example: '10969'
              nullable: true
            address_city:
              example: Berlin
              nullable: true
              $ref: '#/components/schemas/TinyText'
            address_country:
              type: string
              maxLength: 2
              nullable: true
          type: object
        bank_account:
          properties:
            iban:
              description: 'Virtual IBAN provided by Billie'
              nullable: true
              $ref: '#/components/schemas/TinyText'
            bic:
              nullable: true
              $ref: '#/components/schemas/TinyText'
          type: object
        invoice:
          properties:
            invoice_number:
              nullable: true
              $ref: '#/components/schemas/TinyText'
            payout_amount:
              type: number
              format: float
              nullable: true
            outstanding_amount:
              type: number
              format: float
              nullable: true
            fee_amount:
              type: number
              format: float
              nullable: true
            fee_rate:
              type: number
              format: float
              nullable: true
            due_date:
              type: string
              format: date
              example: '2019-03-20'
              nullable: true
          type: object
        debtor_external_data:
          description: 'Data provided in the order creation'
          properties:
            merchant_customer_id:
              example: C-10123456789
              $ref: '#/components/schemas/TinyText'
            name:
              example: 'Billie G.m.b.H.'
              $ref: '#/components/schemas/TinyText'
            address_country:
              type: string
              maxLength: 2
              example: DE
            address_city:
              example: Berlin
              $ref: '#/components/schemas/TinyText'
            address_postal_code:
              type: string
              maxLength: 5
              example: '10969'
            address_street:
              example: 'Charlotten Straße'
              $ref: '#/components/schemas/TinyText'
            address_house:
              example: '4'
              $ref: '#/components/schemas/TinyText'
            industry_sector:
              nullable: true
              $ref: '#/components/schemas/TinyText'
          type: object
        delivery_address:
          type: object
          $ref: '#/components/schemas/CreateOrderAddressRequest'
        billing_address:
          type: object
          $ref: '#/components/schemas/CreateOrderAddressRequest'
        created_at:
          $ref: '#/components/schemas/DateTime'
        shipped_at:
          $ref: '#/components/schemas/DateTime'
      type: object
    ErrorsObject:
      title: 'API Errors'
      properties:
        errors:
          type: array
          items:
            required:
              - title
              - code
            properties:
              title:
                description: 'Error message'
                type: string
              code:
                description: 'Error code'
                type: string
                nullable: true
              source:
                description: 'Property path where the error occurred, if applicable'
                type: string
                nullable: true
            type: object
      type: object
...
  responses:
    HtmlDocument:
      description: 'HTML Document'
      content:
        text/html:
          schema:
            type: string
...
    ServerError:
      description: 'Unexpected Server Error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsObject'
    NotFound:
      description: 'Resource Not Found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsObject'
    Unauthorized:
      description: 'Unauthorized request or invalid credentials'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsObject'
...
    BadRequest:
      description: 'Invalid Request Data'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsObject'
  securitySchemes:
    oauth2:
      type: oauth2
      description: |
        This authentication mechanism is based on the OAuth 2.0 prototocol and requires a Client ID, a Client Secret Key and
        in some cases also a user and password.
        
        With these credentials, the client should request a token using the `POST /api/v1/oauth/token` endpoint,
        which should be used in the `Authorization` header together with the `Bearer` challenge to authenticate all the requests.
        
        > Example: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c`.
      flows:
        password:
          tokenUrl: /oauth/token
          scopes:
            default: default
    apiKey:
      type: apiKey
      description: |
        This authentication mechanism is based on the `X-Api-Key` header. It has been is **deprecated**, since it is
        less secure because API Keys never expire, in comparison to OAuth tokens.
        
        > Example: `X-Api-Key: bf428b05-f80e-4ed0-aa79-e0e7bef274a3`.
      name: X-Api-Key
      in: header
      x-groups:
        - internal
...